# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Fine-Print Guardian API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /watches:
    get:
      operationId: listWatches
      tags: [watches]
      summary: List all active document watches
      responses:
        "200":
          description: List of watches
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Watch"
    post:
      operationId: createWatch
      tags: [watches]
      summary: Start watching a document for changes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/WatchInput"
      responses:
        "201":
          description: Watch created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watch"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /watches/{id}:
    delete:
      operationId: deleteWatch
      tags: [watches]
      summary: Stop watching a document
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /watches/{id}/check:
    post:
      operationId: checkWatch
      tags: [watches]
      summary: Check a watched document for clause changes
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Watch status after check
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Watch"
        "400":
          description: Document not yet analyzed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    Watch:
      type: object
      properties:
        id:
          type: integer
        documentId:
          type: integer
        documentTitle:
          type: string
        active:
          type: boolean
        forwardingAddress:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
        lastCheckedAt:
          type: ["string", "null"]
          format: date-time
        changesDetected:
          type: ["boolean", "null"]
        changeSummary:
          type: ["string", "null"]
      required:
        - id
        - documentId
        - documentTitle
        - active
        - createdAt
    WatchInput:
      type: object
      properties:
        documentId:
          type: integer
      required:
        - documentId
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error