# lib/api-zod/src/generated/types/watchInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Fine-Print Guardian API specification
 * OpenAPI spec version: 0.1.0
 */

export interface WatchInput {
  documentId: number;
}